Fix 'pierone login' on Mac (#82)
Docker for Mac uses the Keychain credentials helper by default, which causes the auth token to be ignored. There was an attempt at fixing it by deleting the `credsStore` key, but the actual key is called `credSstore` ¯\\\_(ツ)\_/¯.
Instead of removing stuff from the user's Docker config, just configure the URL to not use a credentials helper explicitly.